requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=2.2.0=py37hecd8cb5_1
appnope=0.1.2=py37hecd8cb5_1001
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h9ed2024_1
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
blosc=1.21.0=h2842e9f_0
bottleneck=1.3.2=py37hf1fa96c_1
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py37h9ed2024_1003
brunsli=0.1=h23ab428_0
bzip2=1.0.8=h1de35cc_0
c-ares=1.17.1=h9ed2024_0
ca-certificates=2021.10.26=hecd8cb5_2
cairo=1.16.0=h8023c5d_1
certifi=2021.10.8=py37hecd8cb5_0
cffi=1.14.5=py37h2125817_0
cfitsio=3.470=hee0f690_6
charls=2.2.0=h23ab428_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cryptography=35.0.0=py37h2fd3fbb_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37haf1e3a3_0
dask-core=2021.10.0=pyhd3eb1b0_0
debugpy=1.4.1=py37h23ab428_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
ffmpeg=4.0=h01ea3c9_0
fontconfig=2.13.1=ha9ee91d_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
fsspec=2021.10.1=pyhd3eb1b0_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.69.1=hdf23fa2_0
graphite2=1.3.14=h38d11af_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h0a44026_3
idna=3.2=pyhd3eb1b0_0
imagecodecs=2021.8.26=py37ha952a84_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.8.1=py37hecd8cb5_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.4.1=py37hecd8cb5_1
ipython=7.29.0=py37h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.14=h0129ec2_2
jedi=0.18.0=py37hecd8cb5_1
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.9.1=py37hecd8cb5_0
jupyter_server=1.4.1=py37hecd8cb5_0
jupyterlab=3.2.1=pyhd3eb1b0_1
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.8.2=pyhd3eb1b0_0
jxrlib=1.1=haf1e3a3_2
kiwisolver=1.3.1=py37h23ab428_0
krb5=1.19.2=hcd88c3b_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libaec=1.0.4=hb1e8313_1
libcurl=7.78.0=hb8e4fae_0
libcxx=12.0.0=h2f01273_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libnghttp2=1.41.0=h7580e61_2
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.9.0=ha12b0ac_1
libtiff=4.2.0=h87d7836_0
libvpx=1.7.0=h378b8a2_0
libwebp=1.2.0=hacca55c_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
libzopfli=1.0.3=hb1e8313_0
llvm-openmp=12.0.0=h0dcd299_1
locket=0.2.1=py37hecd8cb5_1
lz4-c=1.9.3=h23ab428_1
markupsafe=2.0.1=py37h9ed2024_0
matplotlib-base=3.4.3=py37h0a11d32_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h1de35cc_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py37h9ed2024_0
mkl_fft=1.3.1=py37h4ab4a9b_0
mkl_random=1.2.2=py37hb2f4e1b_0
munkres=1.1.4=py_0
nbclassic=0.2.6=pyhd3eb1b0_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pyhd3eb1b0_0
notebook=6.4.5=py37hecd8cb5_0
numexpr=2.7.3=py37h5873af2_1
numpy=1.21.2=py37h4b4dc7a_0
numpy-base=1.21.2=py37he0bd621_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1l=h9ed2024_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.4=py37h743cdd8_0
pandocfilters=1.4.3=py37hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py37h98e4679_0
pip=21.0.1=py37hecd8cb5_0
pixman=0.40.0=h9ed2024_1
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37h7c891bd_1
pycparser=2.21=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py37hca72f7f_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.11=h88f2d9e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py37haf1e3a3_2
pyyaml=6.0=py37hca72f7f_1
pyzmq=22.2.1=py37h23ab428_1
readline=8.1=h9ed2024_0
requests=2.26.0=pyhd3eb1b0_0
scikit-image=0.18.3=py37hae1ba45_0
scipy=1.7.1=py37h88652d9_2
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=hb1e8313_0
sniffio=1.2.0=py37hecd8cb5_1
sqlite=3.36.0=hce871da_0
terminado=0.9.4=py37hecd8cb5_0
testpath=0.5.0=pyhd3eb1b0_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.11=h7bc2e8c_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.1=py37h9ed2024_0
traitlets=5.1.0=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zfp=0.5.5=h23ab428_6
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
